Update go and geth versions + update test
4/6 tests work. The two tests that are currently failing also failed with the previous version of Geth.

My suspicion is that they worked on an older version of the DB where there was no UNIQUE constraint for header_cids.

I tried to update the `mockCID` and `mockNodeId` for each entry but I kept running into unique contrainst issues. This is something worth looking into down the road.